+++ pkg/sessions/tests/session_store_tests.go
@@ -21,7 +21,7 @@ import (
 // Ginkgo has unpacked the tests.
 // Interfaces have to be wrapped in closures otherwise nil pointers are thrown.
 type testInput struct {
-	cookieOpts            *options.CookieOptions
+	cookieOpts            *options.Cookie
 	ss                    sessionStoreFunc
 	session               *sessionsapi.SessionState
 	request               *http.Request
@@ -38,7 +38,7 @@ type PersistentStoreFastForwardFunc func(time.Duration) error
 
 // NewSessionStoreFunc allows any session store implementation to configure their
 // own session store before each test.
-type NewSessionStoreFunc func(sessionOpts *options.SessionOptions, cookieOpts *options.CookieOptions) (sessionsapi.SessionStore, error)
+type NewSessionStoreFunc func(sessionOpts *options.SessionOptions, cookieOpts *options.Cookie) (sessionsapi.SessionStore, error)
 
 func RunSessionStoreTests(newSS NewSessionStoreFunc, persistentFastForward PersistentStoreFastForwardFunc) {
 	Describe("Session Store Suite", func() {
@@ -62,7 +62,7 @@ func RunSessionStoreTests(newSS NewSessionStoreFunc, persistentFastForward Persi
 			Expect(err).ToNot(HaveOccurred())
 
 			// Set default options in CookieOptions
-			cookieOpts := &options.CookieOptions{
+			cookieOpts := &options.Cookie{
 				Name:     "_oauth2_proxy",
 				Path:     "/",
 				Expire:   time.Duration(168) * time.Hour,
@@ -111,7 +111,7 @@ func RunSessionStoreTests(newSS NewSessionStoreFunc, persistentFastForward Persi
 
 		Context("with non-default options", func() {
 			BeforeEach(func() {
-				input.cookieOpts = &options.CookieOptions{
+				input.cookieOpts = &options.Cookie{
 					Name:     "_cookie_name",
 					Path:     "/path",
 					Expire:   time.Duration(72) * time.Hour,